Fixed issue #4002
Fixed issue #4002
